requirements.txt
Click>=6.0
pyyaml>=3.13
parsel>=1.5.1